laptop/serializers.py
from rest_framework import serializers
from .models import Laptop

class LaptopSerializer(serializers.ModelSerializer):
    class Meta:
        model = Laptop
        fields = ('id','brand','description','owner', 'created_at', 'updated_at')