pyproject.toml
[tool.poetry]
name = "laptop-api"
version = "0.1.0"
description = ""
authors = ["Faisal-Kushha <fkushha89@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.9"
black = "^21.11b1"
flake8 = "^4.0.1"
djangorestframework = "^3.12.4"
pytest = "^6.2.5"
psycopg2-binary = "^2.9.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
